python/requirements.txt
pandas==1.3.5
numpy==1.21.6
matplotlib==3.5.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
statsmodels==0.13.2
pyspellchecker==0.6.2
mplcursors==0.5.3